Fix: eliminate lazy load race condition
When lazily loading images in the demos, create a separate Document to
keep the originals. This eliminates the race condition of removing them
and their source attributes in the transform.

The following alternatives did not work:

- DocumentFragment loads images.
- noscript does not construct a DOM unless JavaScript is disabled.
- The HTML DOMParser extension is unavailable on some Android devices
  such as a Android 4.4.2 KitKat (API 19) Tracfone LG Ultimate 2
  (LGL41C).

As such, removing source attributes from the images is no longer
necessary and may mask improper usage so this is no longer done. It's on
the client to use a separate Document now.