Make document parameters explicit.
Formerly we were using the name 'content' for parameters which could be either a document or a fragment.

This is not strictly needed as there are ways for the caller to enrich any fragment which is to be passed as if it were a document - ie so the fragment can do things such as work with calls to 'createElement' or even 'createDocumentFragment'. Handling such complexity should be the responsibility of the caller.