New: function `collectNearbyReferences()` for extracting adjacents references.
Used on iOS when a reference is tapped to show a panel which lets user drag side to side to see any info for references which were adjacent to the tapped reference. ie: if a lone reference is tapped "[4]" we show a panel with its reference text, but if that reference has adjacent references ie "[3][4][5]" the panel we show can be dragged side to side to see the references text for "[3]" and "[5]".

To use:
Pass `isCitation` the href string of the tapped reference anchor, if it returns true call `collectNearbyReferences` with the document and the tapped anchor element to get extracted references info including a `referencesGroup` array and the index of the selected element `selectedIndex` from that group.

The each item in `referencesGroup` will contain:
 `id` - id of the reference
 `rect` - ClientRect of the reference (so you can do a native overlay dimming everthing but the tapped reference link ie "[4]")
 `text` - the text of the tapped reference link ie "[4]"
 `html` - the html/text of the reference from the actual reference at the bottom of the page