Chore: upgrade Domino v2.0.2 →  2.0.3
Upgrade Domino to v2.0.3. This release includes a fix for URL fragments
which this patch removes a workaround for. Additionally, CSS filters are
now supported by Domino so the CompatibilityTransform test is inverted,
and Element.style.<property> now correctly unconditionally returns an
empty string when undefined so the WidenImage test is updated.

https://github.com/fgnass/domino/issues/127
https://github.com/fgnass/domino/blob/88309f8/CHANGELOG.md#domino-203-12-jul-2018