theme: do not assume white background for dialog icons
A small icon image (lower case i inside a blue circle) in the
infobox of the article Wikipedia:How_to_read_a_color_infobox had an ugly
white background, even though it has transparent background in the image
itself.

While I haven't checked all of them, it seems most icons in
https://commons.wikimedia.org/wiki/Comparison_of_icon_sets#Status
that a dialog icons have a transparent background and don't need a white
background.

The Theme CSS added a white background where it's not needed.
Now the ThemeTransform knows that this icon does not to presume a white
background.