Fix: dependency audit
package.json says Node.js v7.5.0 is used but it does not support
`npm audit fix`. This patch performs a fresh install of node_modules on
v7.5.0 and a `npm audit fix` using a newer version, v11.10.0. This
produced updates in package-lock.json but also upgraded ESLint from
v5.10.0 to v5.14.1 which broke the version of eslint-plugin-html used,
v5.0.0, so that's been upgraded to v5.0.3.

As is characteristic of ESLint, there have been many, many, many
changes. However, I suspect this library will migrate to
eslint-config-wikimedia (T163831) so I have not reviewed them.

Bug: T216396